+++ renderer/src/mesh_layers/text_mesh_layer.rs
@@ -1,4 +1,4 @@
-use crate::GlobalContext;
+use crate::global_context::GlobalContext;
 use crate::geometry_data::TextData;
 use crate::mesh_layers::BaseMeshLayer;
 use crate::mesh_layers::render_data_holder::RenderDataHolder;
