+++ renderer/src/msaa_texture.rs
@@ -1,4 +1,4 @@
-use crate::GlobalContext;
+use crate::global_context::GlobalContext;
 use wgpu::TextureView;
 
 pub struct MultisampledTexture {
