+++ renderer/src/pipelines/mod.rs
@@ -1,7 +1,7 @@
 use crate::mesh::mesh::Mesh;
 use crate::modifier::render_modifier::SpatialData;
 use crate::mesh::positioned_mesh::{MeshInstanceInput, PositionedMesh};
-use crate::GlobalContext;
+use crate::global_context::GlobalContext;
 use cgmath::Vector3;
 use wgpu::{ColorTargetState, DepthStencilState, Device, Label, MultisampleState, PipelineCompilationOptions, PipelineLayout, PrimitiveState, RenderPass, ShaderModule, VertexBufferLayout};
